+++ libs/ngx.command/src/command.ts
@@ -4,34 +4,46 @@ import {
 	tap, filter, switchMap, catchError, finalize, take,
 } from "rxjs";
 import type { CanExecute, ExecuteAsyncFn, ExecuteFn, ICommand } from "./command.model";
-import { computed, DestroyRef, inject, isSignal, signal, type Signal } from "@angular/core";
+import { assertInInjectionContext, computed, DestroyRef, inject, Injector, isSignal, signal, type Signal } from "@angular/core";
 import { toSignal } from "@angular/core/rxjs-interop";
+
+export interface CommandCreateOptions {
+	isAsync: boolean,
+	injector?: Injector;
+}
+
+const COMMAND_ASYNC_DEFAULT_OPTIONS: CommandCreateOptions = { isAsync: true };
 
 /** Creates an async {@link Command}. Must be used within an injection context.
  * NOTE: this auto injects `DestroyRef` and handles auto destroy. {@link ICommand.autoDestroy} should not be used.
  */
-export function createCommandAsync(
+export function commandAsync(
 	execute: ExecuteAsyncFn,
 	canExecute$?: CanExecute,
+	opts?: Omit<CommandCreateOptions, "isAsync">,
 ): Command {
-	return createCommand(execute, canExecute$, true);
+	return command(execute, canExecute$, opts ? { ...opts, ...COMMAND_ASYNC_DEFAULT_OPTIONS } : COMMAND_ASYNC_DEFAULT_OPTIONS);
 }
 
 /** Creates a {@link Command}. Must be used within an injection context.
  * NOTE: this auto injects `DestroyRef` and handles auto destroy. {@link ICommand.autoDestroy} should not be used.
  */
-export function createCommand(
+export function command(
 	execute: ExecuteFn,
 	canExecute$?: CanExecute,
-	isAsync?: boolean,
+	opts?: CommandCreateOptions,
 ): Command {
-	const destroyRef = inject(DestroyRef);
-
+	if (!opts?.injector) {
+		assertInInjectionContext(command);
+	}
+	const injector = opts?.injector ?? inject(Injector);
+	const isAsync = opts?.isAsync ?? false;
+	const destroyRef = injector.get(DestroyRef);
 	const cmd = new Command(execute, canExecute$, isAsync);
 	cmd.autoDestroy = false;
 
 	destroyRef.onDestroy(() => {
-		// console.warn("[createCommandAsync::destroy]");
+		// console.warn("[command::destroy]");
 		cmd.destroy();
 	});
 	return cmd;
@@ -39,6 +51,7 @@ export function createCommand(
 
 /**
  * Command object used to encapsulate information which is needed to perform an action.
+ * @deprecated Use {@link command} or {@link commandAsync} instead.
  */
 export class Command implements ICommand {
 
@@ -87,7 +100,10 @@ export class Command implements ICommand {
 		isAsync?: boolean,
 	) {
 		if (canExecute$) {
-			this._$canExecute = isSignal(canExecute$) ? canExecute$ : toSignal(canExecute$, { initialValue: false });
+			const canExecute = typeof canExecute$ === "function"
+			? computed(canExecute$)
+			: canExecute$;
+			this._$canExecute = isSignal(canExecute) ? canExecute : toSignal(canExecute, { initialValue: false });
 			// this.canExecute$ = combineLatest([
 			// 	this._isExecuting$,
 			// 	isSignal(canExecute$) ? toObservable(canExecute$) : canExecute$
@@ -194,6 +210,7 @@ export class Command implements ICommand {
 /**
  * Async Command object used to encapsulate information which is needed to perform an action,
  * which takes an execute function as Observable/Promise.
+ * @deprecated Use {@link commandAsync} instead.
  */
 export class CommandAsync extends Command {